+++ site/constants/index.ts
@@ -287,3 +287,4 @@ export const MIN_PASSWORD_LENGTH = 8;
 
 export const MAX_FILE_SIZE = 5242880;
 export const ACCEPTED_IMAGE_TYPES = ["image/jpeg", "image/jpg", "image/png"];
+export const HOOTSUITE_URL = "https://platform.hootsuite.com/";
