+++ site/package.json
@@ -91,6 +91,6 @@
         "sass": "^1.57.1",
         "tailwindcss": "^3.2.4",
         "vitest": "^0.29.1",
-        "winston": "^3.8.2"
+        "winston": "^3.9.0"
     }
 }
