+++ site/pages/admin/social-media-accounts.page.tsx
@@ -5,15 +5,126 @@ import { parseCookies } from "nookies";
 import { Fragment, ReactElement, ReactNode } from "react";
 import Table from "../../components/form/Table";
 import { BaseLayout } from "../../components/layout/Layout";
-import { COOKIES_ID_TOKEN, COOKIES_REFRESH_TOKEN } from "../../constants";
+import { COOKIES_ID_TOKEN, COOKIES_REFRESH_TOKEN, HOOTSUITE_URL } from "../../constants";
 import { getParameter, getParametersByPath, putParameter } from "../../data/ssm";
+import { HootsuiteMe, HootsuiteSocialProfiles, HootsuiteToken } from "../../interfaces";
 import { HootsuiteProfiles, SocialMediaAccountsSchema } from "../../schemas/social-media-accounts.schema";
-import { toLowerStartCase } from "../../utils";
+import { hootsuiteTokenCall, toLowerStartCase } from "../../utils";
 import { getSessionWithOrgDetail } from "../../utils/apiUtils/auth";
 
 const title = "Social Media Accounts - Create Transport Disruptions Service";
 const description = "Social Media Accounts page for the Create Transport Disruptions Service";
 
+export const getHootsuiteData = async (
+    ctx: NextPageContext,
+    username: string,
+    orgId: string,
+): Promise<{ clientId: string; userData: SocialMediaAccountsSchema }> => {
+    const cookies = parseCookies(ctx);
+    let clientIdValue = "";
+    let userData: SocialMediaAccountsSchema = [];
+    try {
+        const idToken = cookies[COOKIES_ID_TOKEN];
+        const refreshToken = cookies[COOKIES_REFRESH_TOKEN];
+
+        const clientId = await getParameter(`/social/hootsuite/client_id`);
+        const clientSecret = await getParameter(`/social/hootsuite/client_secret`);
+        if (!clientId || !clientSecret) {
+            throw new Error("clientId and clientSecret must be defined");
+        }
+
+        clientIdValue = clientId.Parameter?.Value || "";
+        const key = `${clientId.Parameter?.Value || ""}:${clientSecret.Parameter?.Value || ""}`;
+
+        const authToken = `Basic ${Buffer.from(key).toString("base64")}`;
+
+        if (idToken && refreshToken)
+            await Promise.all([
+                putParameter(`/${username}/token`, idToken, "SecureString", true),
+                putParameter(`/${username}/refresh-token`, refreshToken, "SecureString", true),
+            ]);
+
+        const tokensByOrganisation = await getParametersByPath(`/social/${orgId}/hootsuite`);
+        const refreshTokens = tokensByOrganisation?.Parameters?.map((token) => {
+            return {
+                value: token.Value,
+                name: token.Name,
+                userId: token?.Name?.split("hootsuite/")[1]?.split("-")[0] ?? "",
+                accountType: startCase(token?.Name?.split("/")[3]) ?? "",
+                addedBy: token?.Name?.split("/")[4]?.replace("_", " ")?.split("-")[1] ?? "",
+            };
+        });
+
+        if (refreshTokens && refreshTokens.length > 0) {
+            await Promise.all(
+                refreshTokens?.map(async (token) => {
+                    const resp = await hootsuiteTokenCall(token.value || "", authToken);
+                    if (resp.ok) {
+                        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                        const tokenResult: HootsuiteToken = await resp.json();
+                        const keys = await getParametersByPath(`/social/${orgId}/hootsuite`);
+
+                        if (!keys || (refreshTokens && keys.Parameters?.length === 0)) {
+                            throw new Error("Refresh token is required to fetch dropdown data");
+                        }
+                        const key: string =
+                            keys.Parameters?.find((rt) => rt.Name?.includes(`${token.userId}`))?.Name || "";
+                        if (!key) {
+                            throw new Error("Refresh token is required to fetch dropdown data");
+                        }
+
+                        await putParameter(key, tokenResult.refresh_token ?? "", "SecureString", true);
+                        const userDetailsResponse = await fetch(`${HOOTSUITE_URL}v1/me`, {
+                            method: "GET",
+                            headers: {
+                                Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
+                            },
+                        });
+                        if (userDetailsResponse.ok) {
+                            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                            const userDetails: HootsuiteMe = await userDetailsResponse.json();
+                            const userInfo = userDetails.data || {};
+
+                            const extraInfo = {
+                                ...userInfo,
+                                accountType: token.accountType || "",
+                                addedBy: token.addedBy || "",
+                                expiresIn: "Never",
+                            } as SocialMediaAccountsSchema[0];
+
+                            const socialProfilesResponse = await fetch(`${HOOTSUITE_URL}v1/socialProfiles`, {
+                                method: "GET",
+                                headers: {
+                                    Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
+                                },
+                            });
+                            if (socialProfilesResponse.ok) {
+                                // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                                const socialProfiles: HootsuiteSocialProfiles = await socialProfilesResponse.json();
+
+                                userData = [
+                                    ...userData,
+                                    {
+                                        ...extraInfo,
+                                        hootsuiteProfiles: (socialProfiles?.data?.map((sp: HootsuiteProfiles[0]) => ({
+                                            type: sp.type,
+                                            socialNetworkId: sp.socialNetworkId,
+                                            id: sp.id,
+                                        })) ?? []) as HootsuiteProfiles,
+                                    },
+                                ] as SocialMediaAccountsSchema;
+                            }
+                        }
+                    }
+                }),
+            );
+        }
+
+        return { clientId: clientIdValue, userData };
+    } catch (e) {
+        throw new Error(`${(e as Error).message}`);
+    }
+};
 export interface SocialMediaAccountsPageProps {
     socialMediaData: SocialMediaAccountsSchema;
     username: string;
@@ -67,7 +178,7 @@ const SocialMediaAccounts = ({ socialMediaData, username, clientId }: SocialMedi
                 <Link
                     className="govuk-button mt-8"
                     data-module="govuk-button"
-                    href={`https://platform.hootsuite.com/oauth2/auth?response_type=code&scope=offline&redirect_uri=http://localhost:3000/api/hootsuite-callback&client_id=${clientId}&state=${username}`}
+                    href={`${HOOTSUITE_URL}oauth2/auth?response_type=code&scope=offline&redirect_uri=http://localhost:3000/api/hootsuite-callback&client_id=${clientId}&state=${username}`}
                 >
                     Connect hootsuite
                 </Link>
@@ -87,120 +198,9 @@ export const getServerSideProps = async (ctx: NextPageContext): Promise<{ props:
         throw new Error("Session data not found");
     }
 
-    let clientIdValue = "";
-    let userData: SocialMediaAccountsSchema = [];
-    const cookies = parseCookies(ctx);
-    try {
-        const idToken = cookies[COOKIES_ID_TOKEN];
-        const refreshToken = cookies[COOKIES_REFRESH_TOKEN];
-
-        const clientId = await getParameter(`/social/hootsuite/client_id`);
-        const clientSecret = await getParameter(`/social/hootsuite/client_secret`);
-        if (!clientId || !clientSecret) {
-            throw new Error("clientId and clientSecret must be defined");
-        }
-
-        clientIdValue = clientId.Parameter?.Value || "";
-        const key = `${clientId.Parameter?.Value || ""}:${clientSecret.Parameter?.Value || ""}`;
-
-        const authToken = `Basic ${Buffer.from(key).toString("base64")}`;
-
-        if (idToken && refreshToken)
-            await Promise.all([
-                putParameter(`/${session.username}/token`, idToken, "SecureString", true),
-                putParameter(`/${session.username}/refresh-token`, refreshToken, "SecureString", true),
-            ]);
-
-        const tokensByOrganisation = await getParametersByPath(`/social/${session.orgId}/hootsuite`);
-        const refreshTokens = await tokensByOrganisation?.Parameters?.map((token) => {
-            return {
-                value: token.Value,
-                name: token.Name,
-                userId: token?.Name?.split("hootsuite/")[1]?.split("-")[0] ?? "",
-                accountType: startCase(token?.Name?.split("/")[3]) ?? "",
-                addedBy: token?.Name?.split("/")[4]?.replace("_", " ")?.split("-")[1] ?? "",
-            };
-        });
-
-        if (refreshTokens && refreshTokens.length > 0) {
-            await Promise.all(
-                refreshTokens?.map(async (token) => {
-                    const resp = await fetch(`https://platform.hootsuite.com/oauth2/token`, {
-                        method: "POST",
-                        body: new URLSearchParams({
-                            grant_type: "refresh_token",
-                            refresh_token: token.value ?? "",
-                        }),
-                        headers: {
-                            "Content-Type": "application/x-www-form-urlencoded",
-                            Authorization: authToken,
-                        },
-                    });
-                    if (resp.ok) {
-                        const tokenResult = await resp.json();
-                        const keys = await getParametersByPath(`/social/${session.orgId}/hootsuite`);
-
-                        if (!keys || (refreshTokens && keys.Parameters?.length === 0)) {
-                            throw new Error("Refresh token is required to fetch dropdown data");
-                        }
-                        const key: string =
-                            keys.Parameters?.find((rt) => rt.Name?.includes(`${token.userId}`))?.Name || "";
-                        if (!key) {
-                            throw new Error("Refresh token is required to fetch dropdown data");
-                        }
-
-                        await putParameter(key, tokenResult.refresh_token ?? "", "SecureString", true);
-                        const userDetailsResponse = await fetch(`https://platform.hootsuite.com/v1/me`, {
-                            method: "GET",
-                            headers: {
-                                Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
-                            },
-                        });
-                        if (userDetailsResponse.ok) {
-                            const userDetails = await userDetailsResponse.json();
-                            const userInfo = userDetails.data || {};
-
-                            const extraInfo = {
-                                ...userInfo,
-                                accountType: token.accountType || "",
-                                addedBy: token.addedBy || "",
-                                expiresIn: "Never",
-                            } as SocialMediaAccountsSchema[0];
-
-                            const socialProfilesResponse = await fetch(
-                                `https://platform.hootsuite.com/v1/socialProfiles`,
-                                {
-                                    method: "GET",
-                                    headers: {
-                                        Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
-                                    },
-                                },
-                            );
-                            if (socialProfilesResponse.ok) {
-                                const socialProfiles = await socialProfilesResponse.json();
-
-                                userData = [
-                                    ...userData,
-                                    {
-                                        ...extraInfo,
-                                        hootsuiteProfiles: (socialProfiles?.data?.map((sp: HootsuiteProfiles[0]) => ({
-                                            type: sp.type,
-                                            socialNetworkId: sp.socialNetworkId,
-                                            id: sp.id,
-                                        })) ?? []) as HootsuiteProfiles,
-                                    },
-                                ] as SocialMediaAccountsSchema;
-                            }
-                        }
-                    }
-                }),
-            );
-        }
-    } catch (e) {
-        throw new Error(`${(e as Error).message}`);
-    }
+    const { clientId, userData } = await getHootsuiteData(ctx, session.username, session.orgId);
     return {
-        props: { socialMediaData: userData, username: session.username, clientId: clientIdValue },
+        props: { socialMediaData: userData, username: session.username, clientId },
     };
 };
