+++ site/pages/create-social-media-post/[disruptionId]/[socialMediaPostIndex].page.tsx
@@ -15,13 +15,12 @@ import {
     REVIEW_DISRUPTION_PAGE_PATH,
 } from "../../../constants";
 import { getDisruptionById } from "../../../data/dynamo";
-import { getParameter, getParametersByPath, putParameter } from "../../../data/ssm";
 import { PageState, ErrorInfo } from "../../../interfaces";
-import { HootsuiteProfiles, SocialMediaAccountsSchema } from "../../../schemas/social-media-accounts.schema";
 import { SocialMediaPost, socialMediaPostSchema } from "../../../schemas/social-media.schema";
 import { destroyCookieOnResponseObject, getPageState } from "../../../utils/apiUtils";
 import { getSession } from "../../../utils/apiUtils/auth";
 import { getStateUpdater, handleBlur } from "../../../utils/formUtils";
+import { getHootsuiteData } from "../../admin/social-media-accounts.page";
 const title = "Create social media message";
 const description = "Create social media message page for the Create Transport Disruptions Service";
 
@@ -241,90 +240,7 @@ export const getServerSideProps = async (ctx: NextPageContext): Promise<{ props:
 
     if (ctx.res) destroyCookieOnResponseObject(COOKIES_SOCIAL_MEDIA_ERRORS, ctx.res);
 
-    const tokensByOrganisation = await getParametersByPath(`/social/${session.orgId}/hootsuite`);
-
-    const refreshTokens = tokensByOrganisation?.Parameters?.map((token) => ({
-        value: token.Value,
-        name: token.Name,
-        userId: token?.Name?.split("hootsuite/")[1].split("-")[0] ?? "",
-    }));
-
-    let userData: SocialMediaAccountsSchema = [];
-    const clientId = await getParameter(`/social/hootsuite/client_id`);
-    const clientSecret = await getParameter(`/social/hootsuite/client_secret`);
-    if (!clientId || !clientSecret) {
-        throw new Error("clientId and clientSecret must be defined");
-    }
-
-    const hootsuiteKey = `${clientId.Parameter?.Value || ""}:${clientSecret.Parameter?.Value || ""}`;
-
-    const authToken = `Basic ${Buffer.from(hootsuiteKey).toString("base64")}`;
-
-    if (refreshTokens) {
-        await Promise.all(
-            refreshTokens?.map(async (token) => {
-                const resp = await fetch(`https://platform.hootsuite.com/oauth2/token`, {
-                    method: "POST",
-                    body: new URLSearchParams({
-                        grant_type: "refresh_token",
-                        refresh_token: token.value ?? "",
-                    }),
-                    headers: {
-                        "Content-Type": "application/x-www-form-urlencoded",
-                        Authorization: authToken,
-                    },
-                });
-
-                if (resp.ok) {
-                    const tokenResult = await resp.json();
-
-                    const keys = await getParametersByPath(`/social/${session.orgId}/hootsuite`);
-
-                    if (!keys || (refreshTokens && keys.Parameters?.length === 0)) {
-                        throw new Error("Refresh token is required to fetch dropdown data");
-                    }
-                    const key: string = keys.Parameters?.find((rt) => rt.Name?.includes(`${token.userId}`))?.Name || "";
-                    if (!key) {
-                        throw new Error("Refresh token is required to fetch dropdown data");
-                    }
-
-                    await putParameter(key, tokenResult.refresh_token ?? "", "SecureString", true);
-                    const userDetailsResponse = await fetch(`https://platform.hootsuite.com/v1/me`, {
-                        method: "GET",
-                        headers: {
-                            Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
-                        },
-                    });
-                    if (userDetailsResponse.ok) {
-                        const userDetails = await userDetailsResponse.json();
-                        const userInfo = userDetails.data || {};
-
-                        const socialProfilesResponse = await fetch(`https://platform.hootsuite.com/v1/socialProfiles`, {
-                            method: "GET",
-                            headers: {
-                                Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
-                            },
-                        });
-                        if (socialProfilesResponse.ok) {
-                            const socialProfiles = await socialProfilesResponse.json();
-
-                            userData = [
-                                ...userData,
-                                {
-                                    ...userInfo,
-                                    hootsuiteProfiles: (socialProfiles?.data?.map((sp: HootsuiteProfiles[0]) => ({
-                                        type: sp.type,
-                                        socialNetworkId: sp.socialNetworkId,
-                                        id: sp.id,
-                                    })) ?? []) as HootsuiteProfiles,
-                                },
-                            ] as SocialMediaAccountsSchema;
-                        }
-                    }
-                }
-            }),
-        );
-    }
+    const { userData } = await getHootsuiteData(ctx, session.username, session.orgId);
 
     const socialAccounts = userData?.map((info) => ({
         value: info.id,
