+++ site/utils/apiUtils/auth.ts
@@ -16,8 +16,6 @@ export const getSession = (req: NextApiRequest | IncomingMessage): Session | nul
     return null;
 };
 
-export const delay = (ms: number) => new Promise((res) => setTimeout(res, ms));
-
 export const getSessionWithOrgDetail = async (
     req: NextApiRequest | IncomingMessage,
 ): Promise<SessionWithOrgDetail | null> => {
