+++ site/utils/apiUtils/index.ts
@@ -1,7 +1,11 @@
+import { SocialMediaPostStatus } from "@create-disruptions-data/shared-ts/enums";
+import dayjs from "dayjs";
 import { NextApiRequest, NextApiResponse } from "next";
 import { parseCookies, setCookie } from "nookies";
 import { z } from "zod";
+import { readFile } from "fs/promises";
 import { IncomingMessage, ServerResponse } from "http";
+import { hootsuiteTokenCall } from "..";
 import {
     COOKIES_POLICY_COOKIE,
     COOKIE_CSRF,
@@ -10,8 +14,12 @@ import {
     DISRUPTION_DETAIL_PAGE_PATH,
     REVIEW_DISRUPTION_PAGE_PATH,
     COOKIES_REFRESH_TOKEN,
+    HOOTSUITE_URL,
 } from "../../constants";
-import { PageState } from "../../interfaces";
+import { upsertSocialMediaPost } from "../../data/dynamo";
+import { getParameter, getParametersByPath, putParameter } from "../../data/ssm";
+import { HootsuiteMedia, HootsuiteMediaStatus, HootsuiteToken, PageState } from "../../interfaces";
+import { SocialMediaPost } from "../../schemas/social-media.schema";
 import logger from "../logger";
 
 export const setCookieOnResponseObject = (
@@ -106,4 +114,196 @@ export const getReturnPage = (req: NextApiRequest) => {
         decodedQueryParam?.includes(DISRUPTION_DETAIL_PAGE_PATH)
         ? queryParam
         : null;
+};
+
+export const delay = (ms: number) => new Promise((res) => setTimeout(res, ms));
+
+export const publishToHootsuite = async (socialMediaPosts: SocialMediaPost[], orgId: string) => {
+    await Promise.all(
+        socialMediaPosts
+            .filter((s) => s.status === SocialMediaPostStatus.pending)
+            .map(async (socialMediaPost) => {
+                const refreshTokens = await getParametersByPath(`/social/${orgId}/hootsuite`);
+
+                if (!refreshTokens || (refreshTokens && refreshTokens.Parameters?.length === 0)) {
+                    await upsertSocialMediaPost(
+                        {
+                            ...socialMediaPost,
+                            status: SocialMediaPostStatus.rejected,
+                        },
+                        orgId,
+                    );
+                    logger.debug("Refresh token is required when creating a social media post");
+                }
+                const refreshToken = refreshTokens.Parameters?.find((rt) =>
+                    rt.Name?.includes(`${socialMediaPost.socialAccount}`),
+                );
+                if (!refreshToken) {
+                    await upsertSocialMediaPost(
+                        {
+                            ...socialMediaPost,
+                            status: SocialMediaPostStatus.rejected,
+                        },
+                        orgId,
+                    );
+                    logger.debug("Refresh token is required when creating a social media post");
+                }
+                const clientId = await getParameter(`/social/hootsuite/client_id`);
+                const clientSecret = await getParameter(`/social/hootsuite/client_secret`);
+                if (!clientId || !clientSecret) {
+                    await upsertSocialMediaPost(
+                        {
+                            ...socialMediaPost,
+                            status: SocialMediaPostStatus.rejected,
+                        },
+                        orgId,
+                    );
+                    logger.debug("clientId and clientSecret must be defined");
+                }
+
+                const credentials = `${clientId.Parameter?.Value || ""}:${clientSecret.Parameter?.Value || ""}`;
+
+                const authToken = `Basic ${Buffer.from(credentials).toString("base64")}`;
+                const responseToken = await hootsuiteTokenCall(refreshToken?.Value ?? "", authToken);
+
+                if (responseToken.ok) {
+                    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                    const tokenResult: HootsuiteToken = await responseToken.json();
+                    const key = refreshToken?.Name || "";
+                    await putParameter(key, tokenResult.refresh_token ?? "", "SecureString", true);
+
+                    let imageLink = { id: "", url: "" };
+                    let canUpload = false;
+
+                    if (socialMediaPost.image) {
+                        const responseImage = await fetch(`${HOOTSUITE_URL}v1/media`, {
+                            method: "POST",
+                            body: JSON.stringify({
+                                sizeBytes: socialMediaPost.image.size,
+                                mimeType: socialMediaPost.image.mimetype,
+                            }),
+                            headers: {
+                                "Content-Type": "application/json",
+                                Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
+                            },
+                        });
+
+                        if (responseImage.ok) {
+                            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                            const image: HootsuiteMedia = await responseImage.json();
+                            const imageContents = await readFile(socialMediaPost.image.filepath || "");
+                            imageLink = { url: image.data.uploadUrl, id: image.data.id };
+
+                            const uploadResponse = await fetch(imageLink.url, {
+                                method: "PUT",
+                                headers: {
+                                    "Content-Type": socialMediaPost.image.mimetype,
+                                },
+                                body: imageContents,
+                            });
+                            if (uploadResponse.ok) {
+                                for (let i = 0; i < 3; i++) {
+                                    const imageStatus = await fetch(`${HOOTSUITE_URL}v1/media/${imageLink.id}`, {
+                                        method: "GET",
+                                        headers: {
+                                            Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
+                                        },
+                                    });
+                                    if (imageStatus.ok) {
+                                        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                                        const imageState: HootsuiteMediaStatus = await imageStatus.json();
+
+                                        if (imageState.data.state === "READY") {
+                                            canUpload = true;
+
+                                            break;
+                                        } else {
+                                            await delay(1000);
+                                        }
+                                    } else {
+                                        await upsertSocialMediaPost(
+                                            {
+                                                ...socialMediaPost,
+                                                status: SocialMediaPostStatus.rejected,
+                                            },
+                                            orgId,
+                                        );
+                                        logger.debug("Cannot retrieve media details from hootsuite");
+                                    }
+                                }
+                                if (!canUpload) {
+                                    await delay(5000);
+                                    canUpload = true;
+                                }
+                            } else {
+                                await upsertSocialMediaPost(
+                                    {
+                                        ...socialMediaPost,
+                                        status: SocialMediaPostStatus.rejected,
+                                    },
+                                    orgId,
+                                );
+                                logger.debug("Cannot upload image to hootsuite");
+                            }
+                        } else {
+                            await upsertSocialMediaPost(
+                                {
+                                    ...socialMediaPost,
+                                    status: SocialMediaPostStatus.rejected,
+                                },
+                                orgId,
+                            );
+                            logger.debug("Cannot retrieve upload url from hootsuite");
+                        }
+                    }
+
+                    const formattedDate = dayjs(
+                        `${socialMediaPost.publishDate} ${socialMediaPost.publishTime}`,
+                        "DD/MM/YYYY HHmm",
+                    ).toISOString();
+
+                    const createSocialPostResponse = await fetch(`${HOOTSUITE_URL}v1/messages`, {
+                        method: "POST",
+                        body: JSON.stringify({
+                            text: socialMediaPost.messageContent,
+                            scheduledSendTime: formattedDate,
+                            socialProfileIds: [socialMediaPost.hootsuiteProfile],
+                            ...(imageLink.id ? { media: [{ id: imageLink.id }] } : {}),
+                        }),
+                        headers: {
+                            "Content-Type": "application/json",
+                            Authorization: `Bearer ${tokenResult.access_token ?? ""}`,
+                        },
+                    });
+
+                    if (!createSocialPostResponse.ok) {
+                        await upsertSocialMediaPost(
+                            {
+                                ...socialMediaPost,
+                                status: SocialMediaPostStatus.rejected,
+                            },
+                            orgId,
+                        );
+                        logger.debug("Failed to create social media post");
+                    } else {
+                        await upsertSocialMediaPost(
+                            {
+                                ...socialMediaPost,
+                                status: SocialMediaPostStatus.successful,
+                            },
+                            orgId,
+                        );
+                    }
+                } else {
+                    await upsertSocialMediaPost(
+                        {
+                            ...socialMediaPost,
+                            status: SocialMediaPostStatus.rejected,
+                        },
+                        orgId,
+                    );
+                    logger.debug("Could not retrieve token from Hootsuite");
+                }
+            }),
+    );
 };
